+++ client/testFixture/includes/boolean_types.h
@@ -1,0 +1,16 @@
+#ifndef BOOLEAN_TYPES_H
+#define BOOLEAN_TYPES_H
+
+#ifndef Boolean
+#define Boolean Integer
+#endif
+
+#ifndef FALSE
+#define FALSE 0
+#endif
+
+#ifndef TRUE
+#define TRUE 1
+#endif
+
+#endif